code restructure and refactoring
