+++ bindings/rust/src/core/helpers.rs
