+++ bindings/rust/src/core/mod.rs
@@ -1,0 +1,9 @@
+mod client;
+mod prover;
+mod verifier;
+pub mod helpers;
+
+pub use client::{ColibriClient, ClientConfig};
+pub use prover::Prover;
+pub use verifier::Verifier;
+pub use helpers::{get_method_support, get_method_type};
