+++ bindings/rust/src/storage/mod.rs
@@ -1,3 +1,5 @@
+pub(crate) mod ffi;
+
 use std::collections::HashMap;
 use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
